+++ bin/varnishd/http1/cache_http1.h
@@ -31,7 +31,7 @@ struct VSC_vbe;
 
 /* cache_http1_fetch.c [V1F] */
 int V1F_SendReq(struct worker *, struct busyobj *, uint64_t *ctr_hdrbytes,
-    uint64_t *ctr_bodybytes, int onlycached, char *addr, char *port);
+    uint64_t *ctr_bodybytes, int onlycached);
 int V1F_FetchRespHdr(struct busyobj *);
 int V1F_Setup_Fetch(struct vfp_ctx *vfc, struct http_conn *htc);