+++ include/tbl/vsl_tags.h
@@ -592,16 +592,11 @@ SLTM(Witness, 0, "Lock order witness records",
 	"Diagnostic recording of locking order.\n"
 )
 
-SLTM(BackendStart, 0, "Backend request start",
-	"Start of backend processing. Logs the backend IP address and port"
-	" number.\n\n"
-	"The format is::\n\n"
-	"\t%s %s\n"
-	"\t|  |\n"
-	"\t|  +- Backend Port number\n"
-	"\t+---- Backend IP4/6 address\n"
-	"\n"
-)
+/*
+ * REL_20190915 remove after VSLng
+ * kept for now for VSL binary compatibility
+ */
+SLTM(BackendStart, 0, "(retired)", "")
 
 SLTM(H2RxHdr, SLT_F_BINARY, "Received HTTP2 frame header",
 	"Binary data"